refactor: added test specific fixtures for readability in snaps
