+++ packages/key-facts/__tests__/shared-key-facts-with-style.native.js
@@ -10,9 +10,9 @@ import {
   replaceTransform,
   stylePrinter
 } from "@times-components/jest-serializer";
-import iterator from "@times-components/test-utils";
+import { iterator } from "@times-components/test-utils";
 import KeyFacts from "../src/key-facts";
-import data from "../fixtures/key-facts.json";
+import data from "../fixtures/key-facts-tests.json";
 
 const { data: { children, attributes } } = data;
