+++ packages/key-facts/fixtures/key-facts-showcase.json
